Update tests according to review
