+++ tests/components/myuplink/test_switch.py
@@ -2,7 +2,9 @@
 
 from unittest.mock import MagicMock
 
-from homeassistant.components.switch import DOMAIN as SWITCH_DOMAIN
+import pytest
+
+from homeassistant.components.switch import DOMAIN as PLATFORM_DOMAIN
 from homeassistant.const import (
     ATTR_ENTITY_ID,
     SERVICE_TURN_OFF,
@@ -12,9 +14,7 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 
-from . import setup_platform
-
-from tests.common import MockConfigEntry
+pytestmark = pytest.mark.parametrize("platforms", [(PLATFORM_DOMAIN,)])
 
 ENTITY_ID = "switch.f730_cu_3x400v_temporary_lux"
 ENTITY_FRIENDLY_NAME = "F730 CU 3x400V Tempo­rary lux"
@@ -25,10 +25,9 @@ async def test_entity_registry(
     hass: HomeAssistant,
     entity_registry: er.EntityRegistry,
     mock_myuplink_client: MagicMock,
-    mock_config_entry: MockConfigEntry,
+    setup_platform,
 ) -> None:
-    """Tests that the devices are registered in the entity registry."""
-    await setup_platform(hass, mock_config_entry, SWITCH_DOMAIN)
+    """Tests that the entities are registered in the entity registry."""
 
     entry = entity_registry.async_get(ENTITY_ID)
     assert entry.unique_id == ENTITY_UID
@@ -37,10 +36,9 @@ async def test_entity_registry(
 async def test_attributes(
     hass: HomeAssistant,
     mock_myuplink_client: MagicMock,
-    mock_config_entry: MockConfigEntry,
+    setup_platform,
 ) -> None:
     """Test the switch attributes are correct."""
-    await setup_platform(hass, mock_config_entry, SWITCH_DOMAIN)
 
     state = hass.states.get(ENTITY_ID)
     assert state.state == STATE_OFF
@@ -53,13 +51,12 @@ async def test_attributes(
 async def test_switch_on(
     hass: HomeAssistant,
     mock_myuplink_client: MagicMock,
-    mock_config_entry: MockConfigEntry,
+    setup_platform,
 ) -> None:
     """Test the switch can be turned on."""
-    await setup_platform(hass, mock_config_entry, SWITCH_DOMAIN)
 
     await hass.services.async_call(
-        SWITCH_DOMAIN, SERVICE_TURN_ON, {ATTR_ENTITY_ID: ENTITY_ID}, blocking=True
+        PLATFORM_DOMAIN, SERVICE_TURN_ON, {ATTR_ENTITY_ID: ENTITY_ID}, blocking=True
     )
     await hass.async_block_till_done()
     mock_myuplink_client.async_set_device_points.assert_called_once()
@@ -67,14 +64,13 @@ async def test_switch_on(
 
 async def test_switch_off(
     hass: HomeAssistant,
-    mock_config_entry: MockConfigEntry,
     mock_myuplink_client: MagicMock,
+    setup_platform,
 ) -> None:
     """Test the switch can be turned on."""
-    await setup_platform(hass, mock_config_entry, SWITCH_DOMAIN)
 
     await hass.services.async_call(
-        SWITCH_DOMAIN, SERVICE_TURN_OFF, {ATTR_ENTITY_ID: ENTITY_ID}, blocking=True
+        PLATFORM_DOMAIN, SERVICE_TURN_OFF, {ATTR_ENTITY_ID: ENTITY_ID}, blocking=True
     )
     await hass.async_block_till_done()
     mock_myuplink_client.async_set_device_points.assert_called_once()
